fix(p5-w1-schema): qa_trace_id BigInteger to match qa_traces.id (Codex MED-2)
Migration 024 had sa.Integer() for qa_trace_id while qa_traces.id is
BigInteger. FK range mismatch flagged by Codex review on T5-02 PR.

ORM already updated to BigInteger by orchestrator inline; this commit
aligns the migration. Tests in ea97e81 verify bigint via
information_schema.columns assertion.

+++ alembic/versions/024_add_llm_usage_ledger_and_cache.py
@@ -27,7 +27,7 @@ def upgrade() -> None:
     op.create_table(
         "llm_usage_ledger",
         sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
-        sa.Column("qa_trace_id", sa.Integer(), nullable=True),
+        sa.Column("qa_trace_id", sa.BigInteger(), nullable=True),
         sa.Column("provider", sa.String(64), nullable=False),
         sa.Column("model", sa.String(128), nullable=False),
         sa.Column("prompt_hash", sa.CHAR(64), nullable=False),
